More styling and some pug edits
Styled input text and submit and also messed with the pug files to include favicon, and some classes/ordering of things. I couldn't figure out how to assign the left and right to sender and receiver on conversation page outside of the list tho... thoughts?

also a bad attempt at project description on new convo page